Switch internal keyval processing to guint32
Keyvals are 32-bit integers (on account of them holding 21-bit unicode
characters + using 25th bit as a flag), use 32-bit integers when
manipulating them.

The gdk_win32_keymap_check_compose32() function is the same as
gdk_win32_keymap_check_compose(), but accepts guint32.
gdk_win32_keymap_check_compose() remains available to avoid breaking
ABI compatibility (obviously, it still can't handle 32-bit integers).

+++ gtk/gtkcomposetable.c
@@ -279,7 +279,7 @@ gtk_compose_list_check_duplicated (GList *compose_list)
 
   for (list = compose_list; list != NULL; list = list->next)
     {
-      static guint16 keysyms[GTK_MAX_COMPOSE_LEN + 1];
+      static guint32 keysyms[GTK_MAX_COMPOSE_LEN + 1];
       int i;
       int n_compose = 0;
       gboolean compose_finish;
@@ -293,7 +293,7 @@ gtk_compose_list_check_duplicated (GList *compose_list)
       for (i = 0; i < GTK_MAX_COMPOSE_LEN + 1; i++)
         {
           gunichar codepoint = compose_data->sequence[i];
-          keysyms[i] = (guint16) codepoint;
+          keysyms[i] = codepoint;
 
           if (codepoint == 0)
             break;
